requirements.txt
# Core dependencies
streamlit>=1.29.0

# LangChain (for tools only, not LLM integration)
langchain>=0.1.0
langchain-core>=0.1.0

# Jira integration
atlassian-python-api>=3.41.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Environment management
python-dotenv>=1.0.0

# HTTP requests for custom LLM client
requests>=2.31.0
urllib3>=1.26.0

# Utilities
pydantic>=2.0.0

# Optional: For advanced analysis
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Development
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0

# Legacy LLM providers (commented out since we use custom client)
# langchain-openai>=0.1.0
# langchain-ollama>=0.1.0
# langchain-community>=0.0.20
